+++ src/components/manage/TextLineEdit/TextLineEdit.jsx
@@ -1,12 +1,12 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { Editor, Node, Transforms, Range, createEditor } from 'slate';
-import { ReactEditor, Editable, Slate, withReact } from 'slate-react';
+/**
+ * A block-building component that implements a single input based on the Slate
+ * Editor. It can render itself as a specified tag.
+ */
+import React from 'react';
 import PropTypes from 'prop-types';
 import { defineMessages, useIntl } from 'react-intl';
-import { usePrevious } from '@plone/volto/helpers';
 import config from '@plone/volto/registry';
-import { P } from '@plone/volto-slate/constants';
-import cx from 'classnames';
+import { TextLineInput } from '@plone/volto/components/manage/Widgets/TextLineWidget';
 
 const messages = defineMessages({
   title: {
@@ -14,6 +14,18 @@ const messages = defineMessages({
     defaultMessage: 'Type the heading…',
   },
 });
+
+export const getFieldValue = ({
+  data,
+  fieldDataName,
+  fieldName = 'title',
+  metadata,
+  properties,
+}) => {
+  return fieldDataName
+    ? data?.[fieldDataName] || ''
+    : (metadata?.[fieldName] ?? properties?.[fieldName] ?? '') || '';
+};
 
 /**
  * A component for inserting an inline textline field for blocks
@@ -26,14 +38,12 @@ export const TextLineEdit = (props) => {
     detached,
     editable,
     index,
-    metadata,
     onAddBlock,
     onChangeBlock,
     onChangeField,
     onFocusNextBlock,
     onFocusPreviousBlock,
     onSelectBlock,
-    properties,
     selected,
     renderTag,
     renderClassName,
@@ -42,89 +52,16 @@ export const TextLineEdit = (props) => {
     placeholder,
   } = props;
 
-  const [editor] = useState(withReact(createEditor()));
-  const [initialValue] = useState([
-    {
-      type: P,
-      children: [
-        {
-          text:
-            data?.[fieldDataName] ||
-            metadata?.[fieldName] ||
-            properties?.[fieldName] ||
-            '',
-        },
-      ],
-    },
-  ]);
-
   const intl = useIntl();
-
-  const prevSelected = usePrevious(selected);
-
-  const text = useMemo(
-    () =>
-      data?.[fieldDataName] ||
-      metadata?.['title'] ||
-      properties?.['title'] ||
-      '',
-    [data, fieldDataName, metadata, properties],
-  );
-
-  const resultantPlaceholder = useMemo(
+  const value = getFieldValue(props);
+  const derivedPlaceholder = React.useMemo(
     () =>
       placeholder || data.placeholder || intl.formatMessage(messages['title']),
     [placeholder, data.placeholder, intl],
   );
-  const disableNewBlocks = useMemo(() => detached, [detached]);
 
-  useEffect(() => {
-    if (!prevSelected && selected) {
-      if (editor.selection && Range.isCollapsed(editor.selection)) {
-        // keep selection
-        ReactEditor.focus(editor);
-      } else {
-        // nothing is selected, move focus to end
-        ReactEditor.focus(editor);
-        Transforms.select(editor, Editor.end(editor, []));
-      }
-    }
-  }, [prevSelected, selected, editor]);
-
-  useEffect(() => {
-    // undo/redo handler
-    const oldText = Node.string(editor);
-    if (oldText !== text) {
-      Transforms.insertText(editor, text, {
-        at: [0, 0],
-      });
-    }
-  }, [editor, text]);
-
-  const handleChange = useCallback(() => {
-    const newText = Node.string(editor);
-    if (newText !== text) {
-      if (fieldDataName) {
-        onChangeBlock(block, {
-          ...data,
-          [fieldDataName]: newText,
-        });
-      } else {
-        onChangeField(fieldName, newText);
-      }
-    }
-  }, [
-    data,
-    block,
-    editor,
-    fieldDataName,
-    fieldName,
-    onChangeField,
-    onChangeBlock,
-    text,
-  ]);
-
-  const handleKeyDown = useCallback(
+  const disableNewBlocks = React.useMemo(() => detached, [detached]);
+  const handleKeyDown = React.useCallback(
     (ev) => {
       if (ev.key === 'Return' || ev.key === 'Enter') {
         ev.preventDefault();
@@ -153,38 +90,50 @@ export const TextLineEdit = (props) => {
     ],
   );
 
-  const handleFocus = useCallback(() => {
+  const handleFocus = React.useCallback(() => {
     onSelectBlock(block);
   }, [block, onSelectBlock]);
 
-  const RenderTag = renderTag || 'h1';
-
-  const renderElement = useCallback(
-    ({ attributes, children }) => {
-      return (
-        <RenderTag {...attributes} className={cx(renderClassName)}>
-          {children}
-        </RenderTag>
-      );
+  const handleChange = React.useCallback(
+    (newValue) => {
+      if (newValue !== value) {
+        if (fieldDataName) {
+          onChangeBlock(block, {
+            ...data,
+            [fieldDataName]: newValue,
+          });
+        } else {
+          onChangeField(fieldName, newValue);
+        }
+      }
     },
-    [renderClassName],
+    [
+      data,
+      block,
+      fieldDataName,
+      fieldName,
+      onChangeField,
+      onChangeBlock,
+      value,
+    ],
   );
 
-  if (typeof window.__SERVER__ !== 'undefined') {
-    return <div />;
-  }
-  return (
-    <Slate editor={editor} onChange={handleChange} value={initialValue}>
-      <Editable
-        readOnly={!editable}
-        onKeyDown={handleKeyDown}
-        placeholder={resultantPlaceholder}
-        renderElement={renderElement}
-        onFocus={handleFocus}
-        aria-multiline="false"
-      ></Editable>
-    </Slate>
-  );
+  return typeof window.__SERVER__ === 'undefined' ? (
+    <TextLineInput
+      readOnly={!editable}
+      placeholder={derivedPlaceholder}
+      value={value}
+      onChange={handleChange}
+      focus={selected}
+      as={renderTag}
+      className={renderClassName}
+      getInputProps={() => ({
+        onKeyDown: handleKeyDown,
+        onFocus: handleFocus,
+        // focus: selected ? 'true' : undefined,
+      })}
+    />
+  ) : null;
 };
 
 TextLineEdit.propTypes = {